Assume role with external ID support
Most likely if using assume-role then an external ID should be
specified, Amazon includes this as a mitigation for the confused deputy
problem. In the context of logstash this may be the following:

  - Alice trusts LoggingProviderCorp to read logs from their AWS account
  - Mallory signs up to LoggingProviderCorp and asks them to read logs
    out of Alice's account.

Because the trust relationship is between Alice and LoggingProviderCorp
it can be taken advantage of by Mallory, making LoggingProviderCorp a
so-called "confused deputy".

External ID's are used to achive the following:

  - Alice trusts LoggingProviderCorp to read logs from their AWS account
    and assigns Alice an "external_id" it will include on any request
    accessing Alice's account.
  - Mallory signs up to LoggingProviderCorp and asks them to read logs
    out of Alice's account, because LoggingProviderCorp will not pass
    the same "external_id" the requests are blocked by an AWS policy.